Revert automatic container detection mechanism
Turns out it's more complicated to start the IBIS through Spring SPI..
