+++ core/src/main/java/nl/nn/adapterframework/lifecycle/IbisApplicationServlet.java
@@ -69,6 +69,10 @@ public class IbisApplicationServlet extends CXFServlet {
 	public void init(ServletConfig config) throws ServletException {
 		this.currentConfig = config;
 		servletContext = config.getServletContext();
+
+		checkAndCorrectLegacyServerTypes();
+		determineApplicationServerType();
+
 		AppConstants appConstants = AppConstants.getInstance();
 		String realPath = servletContext.getRealPath("/");
 		if (realPath != null) {
@@ -162,6 +166,55 @@ public class IbisApplicationServlet extends CXFServlet {
 			context.setAttribute("javax.servlet.context.tempdir",tempDirFile);
 		} catch (Exception e) {
 			log.error("Could not set servlet context attribute 'javax.servlet.context.tempdir' to value of ${upload.dir}",e);
+		}
+	}
+
+	private void checkAndCorrectLegacyServerTypes() {
+		//In case the property is explicitly set with an unsupported value, E.g. 'applName + number'
+		String applicationServerType = System.getProperty(AppConstants.APPLICATION_SERVER_TYPE_PROPERTY);
+		if (StringUtils.isNotEmpty(applicationServerType)) {
+			if (applicationServerType.equalsIgnoreCase("WAS5") || applicationServerType.equalsIgnoreCase("WAS6")) {
+				log("interpeting value ["+applicationServerType+"] of property ["+AppConstants.APPLICATION_SERVER_TYPE_PROPERTY+"] as [WAS]");
+				System.setProperty(AppConstants.APPLICATION_SERVER_TYPE_PROPERTY, "WAS");
+			} else if (applicationServerType.equalsIgnoreCase("TOMCAT6")) {
+				log("interpeting value ["+applicationServerType+"] of property ["+AppConstants.APPLICATION_SERVER_TYPE_PROPERTY+"] as [TOMCAT]");
+				System.setProperty(AppConstants.APPLICATION_SERVER_TYPE_PROPERTY, "TOMCAT");
+			}
+		}
+	}
+
+	private void determineApplicationServerType() {
+		String serverInfo = servletContext.getServerInfo();
+		String defaultApplicationServerType = null;
+		if (StringUtils.containsIgnoreCase(serverInfo, "WebSphere Liberty")) {
+			defaultApplicationServerType = "WLP";
+		} else if (StringUtils.containsIgnoreCase(serverInfo, "WebSphere")) {
+			defaultApplicationServerType = "WAS";
+		} else if (StringUtils.containsIgnoreCase(serverInfo, "Tomcat")) {
+			defaultApplicationServerType = "TOMCAT";
+		} else if (StringUtils.containsIgnoreCase(serverInfo, "JBoss")) {
+			defaultApplicationServerType = "JBOSS";
+		} else if (StringUtils.containsIgnoreCase(serverInfo, "WildFly")) {
+			defaultApplicationServerType = "JBOSS";
+		} else if (StringUtils.containsIgnoreCase(serverInfo, "jetty")) {
+			String javaHome = System.getProperty("java.home");
+			if (StringUtils.containsIgnoreCase(javaHome, "tibco")) {
+				defaultApplicationServerType = "TIBCOAMX";
+			} else {
+				defaultApplicationServerType = "JETTYMVN";
+			}
+		} else {
+			defaultApplicationServerType = "TOMCAT";
+			log("unknown server info ["+serverInfo+"] default application server type could not be determined, TOMCAT will be used as default value");
+		}
+
+		//has it explicitly been set? if not, set the property
+		String serverType = System.getProperty(AppConstants.APPLICATION_SERVER_TYPE_PROPERTY);
+		if (defaultApplicationServerType.equals(serverType)) { //and is it the same as the automatically detected version?
+			log("property ["+AppConstants.APPLICATION_SERVER_TYPE_PROPERTY+"] already has a default value ["+defaultApplicationServerType+"]");
+		}
+		else if (StringUtils.isEmpty(serverType)) { //or has it not been set?
+			System.setProperty(AppConstants.APPLICATION_SERVER_TYPE_PROPERTY, defaultApplicationServerType);
 		}
 	}
 }
